did first part of homework number five
