+++ app.js
@@ -3,16 +3,7 @@ const path = require("path");
 const logger = require("morgan");
 const cors = require("cors");
 const helmet = require("helmet");
-const rateLimit = require("express-rate-limit");
-
-// Продолжи создание REST API для работы с коллекцией контактов.
-// Добавь возможность загрузки аватарки пользователя через Multer.
-
-// Создай папку public для раздачи статики. В этой папке сделай папку images.
-//  Настрой Express на раздачу статических файлов из папки public.
-// Положи любое изображение в папку public/images и проверь что раздача статики работает.
-// При переходе по такому URL браузер отобразит изображение.
-// http://locahost:<порт>/images/avatar-upload.png
+const apiLimiter = require("./helpers/apiLimiter");
 
 const contactsRouter = require("./routes/api/contacts");
 const authenticationRoute = require("./routes/api/user");
@@ -22,19 +13,6 @@ const FOLDER_IMAGES = process.env.DIR_IMAGES;
 app.use(express.static(path.join(__dirname, FOLDER_IMAGES)));
 console.log(path.join(__dirname, FOLDER_IMAGES));
 const formatsLogger = app.get("env") === "development" ? "dev" : "short";
-
-const apiLimiter = rateLimit({
-  windowMs: 15 * 60 * 1000, // 15 minutes
-  max: 200,
-  handler: (req, res, next) => {
-    return res.status(400).json({
-      status: "error",
-      code: 400,
-      data: "Bad request",
-      message: "Too many requests, please try again later.",
-    });
-  },
-});
 
 app.use(helmet());
 app.use(logger(formatsLogger));
