+++ helpers/upload.js
@@ -4,10 +4,10 @@ require("dotenv").config();
 const DIR_UPLOAD = path.join(process.cwd(), process.env.DIR_UPLOAD);
 
 const storage = multer.diskStorage({
-  destination: function (req, file, cb) {
+  destination: function (_req, _file, cb) {
     cb(null, DIR_UPLOAD);
   },
-  filename: function (req, file, cb) {
+  filename: function (_req, file, cb) {
     cb(null, file.originalname);
   },
 });
@@ -15,7 +15,7 @@ const storage = multer.diskStorage({
 const upload = multer({
   storage: storage,
   limits: { fileSize: 2000000 },
-  fileFilter: (req, file, cb) => {
+  fileFilter: (_req, file, cb) => {
     if (file.mimetype.includes("image")) {
       cb(null, true);
       return;
